backend/todo/serializers.py
# serializers.py

from rest_framework import serializers
from .models import Task, Category, Context

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = "__all__"

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = "__all__"

class ContextSerializer(serializers.ModelSerializer):
    class Meta:
        model = Context
        fields = ['id', 'content', 'timestamp', 'source_type']
